refactored grakn engine
