+++ grakn-dist/src/grakn
@@ -285,9 +285,9 @@ queue_wipe_all_data() {
 grakn_start_process() {
   local status=
   java -cp "${CLASSPATH}" -Dgrakn.dir="${GRAKN_HOME}/services" -Dgrakn.conf="${GRAKN_CONFIG}" \
-    ai.grakn.engine.GraknEngineServer > /dev/null 2>&1 &
+    ai.grakn.engine.GraknApplication > /dev/null 2>&1 &
   status=$?
-  pid=`ps -ef | grep GraknEngineServer | grep -v grep | awk '{ print $2}'`
+  pid=`ps -ef | grep GraknApplication | grep -v grep | awk '{ print $2}'`
   echo $pid > "${GRAKN_PID}"
   return $status
 }
