+++ grakn-engine/pom.xml
@@ -212,6 +212,18 @@
     <build>
         <plugins>
             <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>3.0.2</version>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
                 <version>1.3.4</version>
                 <groupId>org.raml.plugins</groupId>
                 <artifactId>jaxrs-raml-maven-plugin</artifactId>
