+++ grakn-test/test-integration/pom.xml
@@ -45,6 +45,19 @@
         </dependency>
         <dependency>
             <groupId>ai.grakn</groupId>
+            <artifactId>grakn-engine</artifactId>
+            <version>${project.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.slf4j</groupId>
+                    <artifactId>slf4j-log4j12</artifactId>
+                </exclusion>
+            </exclusions>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>ai.grakn</groupId>
             <artifactId>grakn-graql-shell</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
