+++ grakn-test/test-integration/src/test/java/ai/grakn/test/DistributionContext.java
@@ -20,6 +20,7 @@ package ai.grakn.test;
 
 import ai.grakn.GraknSystemProperty;
 import ai.grakn.client.Client;
+import ai.grakn.engine.GraknApplication;
 import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.engine.util.SimpleURI;
 import ai.grakn.redismock.RedisServer;
@@ -149,7 +150,7 @@ public class DistributionContext extends ExternalResource {
                 "-cp", getClassPath(),
                 "-Dgrakn.dir=" + DIST_DIRECTORY + "/services",
                 "-Dgrakn.conf=" + propertiesFile.getAbsolutePath(),
-                "ai.grakn.engine.GraknEngineServer", "&"};
+                GraknApplication.class.getName(), "&"};
 
         // Start process
         ProcessBuilder processBuilder = new ProcessBuilder(commands);
