+++ grakn-test/test-integration/src/test/java/ai/grakn/test/GraknTestEngineSetup.java
@@ -19,6 +19,7 @@ package ai.grakn.test;
 
 import ai.grakn.GraknTx;
 import ai.grakn.GraknTxType;
+import ai.grakn.engine.EngineTestHelper;
 import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.engine.GraknEngineServer;
 import ai.grakn.engine.SystemKeyspace;
@@ -33,7 +34,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 import static ai.grakn.engine.GraknEngineConfig.JWT_SECRET_PROPERTY;
-import static ai.grakn.engine.GraknEngineServer.configureSpark;
+import static ai.grakn.engine.HttpHandler.configureSpark;
 import static ai.grakn.graql.Graql.var;
 
 /**
@@ -84,7 +85,7 @@ public abstract class GraknTestEngineSetup {
 
         // start engine
         setRestAssuredUri(config);
-        GraknEngineServer server = GraknEngineServer.create(config);
+        GraknEngineServer server = EngineTestHelper.graknEngineServer(config);
         server.start();
 
         LOG.info("engine started.");
