+++ grakn-test/test-integration/src/test/java/ai/grakn/test/engine/GraknEngineServerTest.java
@@ -19,6 +19,7 @@
 package ai.grakn.test.engine;
 
 import ai.grakn.Keyspace;
+import ai.grakn.engine.EngineTestHelper;
 import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.engine.GraknEngineServer;
 import ai.grakn.engine.data.RedisWrapper;
@@ -83,7 +84,7 @@ public class GraknEngineServerTest {
         mock.server().start();
 
         // Start Engine
-        try (GraknEngineServer server = GraknEngineServer.create(conf)) {
+        try (GraknEngineServer server = EngineTestHelper.graknEngineServer(conf)) {
             server.start();
             assertThat(server.getTaskManager(), instanceOf(RedisTaskManager.class));
         }
@@ -97,7 +98,7 @@ public class GraknEngineServerTest {
         RedisServer redisServer = MockRedisRule.create(new SimpleURI(conf.getProperty(REDIS_HOST)).getPort()).server();
         redisServer.start();
 
-        try (GraknEngineServer server = GraknEngineServer.create(conf)) {
+        try (GraknEngineServer server = EngineTestHelper.graknEngineServer(conf)) {
             server.start();
             assertNotNull(server.factory().systemKeyspace());
 
@@ -115,7 +116,7 @@ public class GraknEngineServerTest {
     public void whenEngineServerIsStartedTheFirstTime_TheVersionIsRecordedInRedis() {
         when(jedis.get(VERSION_KEY)).thenReturn(null);
 
-        try (GraknEngineServer server = GraknEngineServer.create(conf, redisWrapper)) {
+        try (GraknEngineServer server = EngineTestHelper.graknEngineServer(conf, redisWrapper)) {
             server.start();
         }
 
@@ -126,7 +127,7 @@ public class GraknEngineServerTest {
     public void whenEngineServerIsStartedASecondTime_TheVersionIsNotChanged() {
         when(jedis.get(VERSION_KEY)).thenReturn(GraknVersion.VERSION);
 
-        try (GraknEngineServer server = GraknEngineServer.create(conf, redisWrapper)) {
+        try (GraknEngineServer server = EngineTestHelper.graknEngineServer(conf, redisWrapper)) {
             server.start();
         }
 
@@ -138,7 +139,7 @@ public class GraknEngineServerTest {
         when(jedis.get(VERSION_KEY)).thenReturn(OLD_VERSION);
         stdout.enableLog();
 
-        try (GraknEngineServer server = GraknEngineServer.create(conf, redisWrapper)) {
+        try (GraknEngineServer server = EngineTestHelper.graknEngineServer(conf, redisWrapper)) {
             server.start();
         }
 
@@ -150,7 +151,7 @@ public class GraknEngineServerTest {
     public void whenEngineServerIsStartedWithDifferentVersion_TheVersionIsNotChanged() {
         when(jedis.get(VERSION_KEY)).thenReturn(OLD_VERSION);
 
-        try (GraknEngineServer server = GraknEngineServer.create(conf, redisWrapper)) {
+        try (GraknEngineServer server = EngineTestHelper.graknEngineServer(conf, redisWrapper)) {
             server.start();
         }
